separte view configuration from OpenXRContext
